requirements.txt
pandas
numpy
scikit-learn
pytorch
tqdm